Restructure the modules and guard `Debug` under a separate `debug` feature
