+++ impl/Cargo.toml
@@ -32,7 +32,7 @@ unicode-xid = { version = "0.2.2", optional = true }
 rustc_version = { version = "0.4", optional = true }
 
 [dev-dependencies]
-derive_more = { path = "..", features = ["add", "display", "from_str", "not", "try_into"] }
+derive_more = { path = "..", features = ["add", "debug", "display", "from_str", "not", "try_into"] }
 itertools = "0.10.5"
 
 [badges]
@@ -49,6 +49,7 @@ add = []
 as_mut = []
 as_ref = []
 constructor = []
+debug = ["syn/extra-traits", "dep:unicode-xid"]
 deref = []
 deref_mut = []
 display = ["syn/extra-traits", "dep:unicode-xid"]
